調整資料夾，暫時取消 prisma
