+++ .eslintrc.js
@@ -1,17 +1,27 @@
-/* eslint-disable no-tabs */
 module.exports = {
-  root: true,
-  env: {
-    browser: true,
-    node: true
-  },
-  extends: [
-    '@nuxtjs/eslint-config-typescript',
-    'plugin:nuxt/recommended',
-    'plugin:vue/vue3-recommended'
-  ],
-  plugins: [
-  ],
-  // add your custom rules here
-  rules: {}
-}
+	root: true,
+	env: {
+		browser: true,
+		node: true,
+		es6: true,
+	},
+	extends: [
+		'eslint:recommended',
+		'plugin:vue/vue3-recommended',
+		'plugin:@typescript-eslint/recommended',
+		'plugin:prettier/recommended',
+		'plugin:nuxt/recommended',
+	],
+	parser: 'vue-eslint-parser',
+	parserOptions: {
+		ecmaVersion: 13,
+		parser: '@typescript-eslint/parser',
+		sourceType: 'module',
+	},
+	rules: {
+		'comma-dangle': 'off',
+		'vue/multi-word-component-names': 0,
+		'space-before-functino-paren': 0,
+		'prettier/prettier': ['error', { endOfLine: 'auto' }],
+	},
+};
