+++ package.json
@@ -1,32 +1,32 @@
 {
-  "name": "volleyball-score-sheet",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "nuxt",
-    "build": "nuxt build",
-    "start": "nuxt start",
-    "generate": "nuxt generate",
-    "lint:js": "eslint --ext \".js,.ts,.vue\" --ignore-path .gitignore .",
-    "lint": "npm run lint:js",
-    "lintfix": "npm run lint:js -- --fix"
-  },
-  "dependencies": {
-    "core-js": "^3.25.3",
-    "element-ui": "^2.15.10",
-    "nuxt": "^2.15.8",
-    "vue": "^2.7.10",
-    "vue-server-renderer": "^2.7.10",
-    "vue-template-compiler": "^2.7.10"
-  },
-  "devDependencies": {
-    "@babel/eslint-parser": "^7.19.1",
-    "@nuxt/types": "^2.15.8",
-    "@nuxt/typescript-build": "^2.1.0",
-    "@nuxtjs/eslint-config-typescript": "^11.0.0",
-    "@nuxtjs/eslint-module": "^3.1.0",
-    "eslint": "^8.36.0",
-    "eslint-plugin-nuxt": "^4.0.0",
-    "eslint-plugin-vue": "^9.10.0"
-  }
+	"name": "nuxt-app",
+	"private": true,
+	"scripts": {
+		"build": "nuxt build",
+		"dev": "nuxt dev",
+		"generate": "nuxt generate",
+		"preview": "nuxt preview",
+		"postinstall": "nuxt prepare"
+	},
+	"devDependencies": {
+		"@element-plus/icons-vue": "^2.1.0",
+		"@element-plus/nuxt": "^1.0.4",
+		"@nuxtjs/color-mode": "^3.2.0",
+		"element-plus": "^2.3.1",
+		"nuxt": "^3.3.2",
+		"prisma": "^4.12.0",
+		"sass": "^1.60.0",
+		"eslint": "^8.9.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-prettier": "^4.0.0",
+		"eslint-plugin-vue": "^8.4.1",
+		"@typescript-eslint/eslint-plugin": "^5.14.0",
+		"@typescript-eslint/parser": "^5.14.0",
+		"vue-eslint-parser": "^8.3.0"
+	},
+	"dependencies": {
+		"@pinia/nuxt": "^0.4.7",
+		"@prisma/client": "^4.12.0",
+		"pinia": "^2.0.33"
+	}
 }
