+++ pages/games/index.vue
