+++ pages/new-game/index.vue
