+++ pages/new-game/set-list/[id].vue
