+++ pages/players/index.vue
