+++ src/components/ClickBox.vue
