--- volleyball-score-sheet/.gitignore
@@ -1,9 +1,0 @@
-node_modules
-*.log*
-.nuxt
-.nitro
-.cache
-.output
-.env
-dist
-.DS_Store
