--- volleyball-score-sheet/README.md
@@ -1,7 +1,0 @@
-# 比賽即時計分與數據呈現
-
-## Mind Map
-
-https://xmind.works/3ePjxbVc
-
-![Mind Map](public\mind-map.png)
