--- volleyball-score-sheet/package.json
@@ -1,32 +1,0 @@
-{
-	"name": "nuxt-app",
-	"private": true,
-	"scripts": {
-		"build": "nuxt build",
-		"dev": "nuxt dev",
-		"generate": "nuxt generate",
-		"preview": "nuxt preview",
-		"postinstall": "nuxt prepare"
-	},
-	"devDependencies": {
-		"@element-plus/icons-vue": "^2.1.0",
-		"@element-plus/nuxt": "^1.0.4",
-		"@nuxtjs/color-mode": "^3.2.0",
-		"element-plus": "^2.3.1",
-		"nuxt": "^3.3.2",
-		"prisma": "^4.12.0",
-		"sass": "^1.60.0",
-		"eslint": "^8.9.0",
-		"eslint-config-prettier": "^8.5.0",
-		"eslint-plugin-prettier": "^4.0.0",
-		"eslint-plugin-vue": "^8.4.1",
-		"@typescript-eslint/eslint-plugin": "^5.14.0",
-		"@typescript-eslint/parser": "^5.14.0",
-		"vue-eslint-parser": "^8.3.0"
-	},
-	"dependencies": {
-		"@pinia/nuxt": "^0.4.7",
-		"@prisma/client": "^4.12.0",
-		"pinia": "^2.0.33"
-	}
-}
